frr-reload: cannot enter sub-context for defining key chain keys
when defining key chains, frr-reload does not get into correct sub-context after entering
a key index. this makes frr ignore the 'key #' sub-context when the 'key-string' is entered and those the
key is added to the incorrect context and failes to be added.
This commit fixes the issue by allowing frr-reload to enter a new sub-context on the 'key #' keyword of 'key chain'

Signed-off-by: Runar Borge <runar@borge.nu>